add drainer config
