+++ integration/utils/error.go
@@ -1,0 +1,12 @@
+package utils
+
+import "github.com/giantswarm/microerror"
+
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailed",
+}
+
+// IsExecutionFailedError asserts executionFailedError.
+func IsExecutionFailedError(err error) bool {
+	return microerror.Cause(err) == executionFailedError
+}
